requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain>=0.0.188
langchain-community>=0.0.1
openai>=0.27.7
requests>=2.31.0
pandas>=1.5.3
sqlalchemy>=2.0.19
